rename metahandler file
