+++ src/ravelights/interface/restapi.py
@@ -9,7 +9,7 @@ from flask_restful import Api, Resource, fields, marshal_with  # type: ignore
 from flask_socketio import SocketIO, emit  # type: ignore
 from loguru import logger
 from ravelights.core.eventhandler import EventHandler
-from ravelights.core.methandler import MetaHandler
+from ravelights.core.meta_handler import MetaHandler
 from ravelights.core.patternscheduler import PatternScheduler
 from ravelights.core.settings import Settings
